+++ mainapp/management/commands/delete_file.py
@@ -3,14 +3,14 @@ Manually delete a file from minio
 """
 from django.core.management.base import BaseCommand
 
-from mainapp.functions.minio import minio_client, minio_file_bucket
+from mainapp.models import File
 
 
 class Command(BaseCommand):
-    help = "Manually delete a file from minio"
+    help = "Manually delete a file so that it won't be reimported"
 
     def add_arguments(self, parser):
         parser.add_argument("id", type=int)
 
     def handle(self, *args, **options):
-        minio_client().remove_object(minio_file_bucket, str(options["id"]))
+        File.objects.get(pk=options["id"]).manually_delete()